// Angular/src/app/pages/interfaces/atualiza-cotacao-json.ts
export interface IitensCotacao{
    product: string;
    approve: string;
    productCode: any;
    quotationId: string;
    userId: any;
    supplierId: any;
    paymentTermsId: any;
    unit: string;
    quantity: string;
    quotationQuantity: string;
    validDate: string;
    necessityDate: string;
    valueICMS: string;
    valueIPI: string;
    unitValue: string;
    totalQuotation: string;
    observation: string;
    observationSupplier: string;
    bestValue: string;
    shipping: string;
    shippingValue: string;
    attachment: any;
}